Revert "ES6 Transpiling"
This reverts commit 384be9ad34b904e721f24c6a2e33bb2e89842df6.
