--- .gitignore
@@ -1,3 +1,0 @@
-Babel/node_modules
-Babel/package-lock.json
-Babel/Build
