--- Babel/package.json
@@ -1,13 +1,0 @@
-{
-    "devDependencies": {
-        "@babel/cli": "^7.18.6",
-        "@babel/core": "^7.18.6",
-        "@babel/preset-env": "^7.18.6",
-        "babelify": "^10.0.0",
-        "browserify": "^17.0.0",
-        "esmify": "^2.1.1"
-    },
-    "dependencies": {
-        "core-js": "^3.23.5"
-    }
-}
